+++ W2_AssignmentScene.cpp
@@ -228,7 +228,7 @@ void W2_AssignmentScene::Draw() const
 
 void W2_AssignmentScene::onTrigger(PxTriggerPair* pairs, PxU32 count)
 {
-	for (auto i{ 0 }; i < count; ++i)
+	for (auto i{ 0 }; i < static_cast<int>(count); ++i)
 	{
 		//ignore delted shapes
 		if (pairs[i].flags & (PxTriggerPairFlag::eREMOVED_SHAPE_TRIGGER | PxTriggerPairFlag::eREMOVED_SHAPE_OTHER))
